requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
astropy>=6.0
python-dotenv>=1.0
